Move BaseStepContext to the top level
